Change strstr() to Jody's two-way implementation
